fix: set `NONOPT_FSST` for 32-bit platforms (fixes #37)
The optimized code paths for (at least) `count1GetNext()` don't work
correctly on 32-bit platforms (at least on ARM, with gcc), presumably
due to using `__builtin_ctzl()` on a 64-bit value. This resulted in
`buildSymbolTable()` being stuck in an endless loop.

This change simply disables the optimized code paths if a 32-bit
platform is detected.